+++ library/src/main/java/ru/bartwell/exfilepicker/ui/adapter/holder/FileFilesListHolder.java
@@ -1,21 +1,19 @@
 package ru.bartwell.exfilepicker.ui.adapter.holder;
 
 import android.content.Context;
-import android.graphics.Bitmap;
-import android.os.Build;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v7.widget.AppCompatImageView;
 import android.support.v7.widget.AppCompatTextView;
 import android.view.View;
 
+import com.bumptech.glide.Glide;
+
 import java.io.File;
 
 import ru.bartwell.exfilepicker.R;
 import ru.bartwell.exfilepicker.ui.callback.OnListItemClickListener;
 import ru.bartwell.exfilepicker.utils.Utils;
-import ru.bartwell.exfilepicker.utils.thumbnail.ThumbnailCache;
-import ru.bartwell.exfilepicker.utils.thumbnail.ThumbnailLoader;
 
 /**
  * Created by BArtWell on 26.02.2017.
@@ -42,15 +40,10 @@ public class FileFilesListHolder extends BaseFilesListHolder {
             mFileSize.setVisibility(View.VISIBLE);
             mFileSize.setText(Utils.getHumanReadableFileSize(mContext, file.length()));
         }
-        if (ThumbnailLoader.canHasThumbnail(file.getName())) {
-            Bitmap bitmap = ThumbnailCache.getInstance().get(file.getAbsolutePath());
-            if (bitmap == null) {
-                new ThumbnailLoader(mContext, mThumbnail).execute(file);
-            } else {
-                mThumbnail.setImageBitmap(bitmap);
-            }
-        } else {
-            mThumbnail.setImageResource(R.drawable.efp__ic_file);
-        }
+
+        Glide.with(mContext)
+                .load(file)
+                .error(R.drawable.efp__ic_file)
+                .into(mThumbnail);
     }
 }